Update startercode.m
I added a lot of comments to ensure that we know what each variable type is, each function is.

The container map might be useless because bagofngrams might have already created a container map. but once you have the text files then I can try and see if it's useful or not.